Fold the rust-lib crate into the rust crate
