+++ Cargo.toml
@@ -40,7 +40,6 @@ wit-bindgen-rust = { path = "crates/rust", version = "0.12.0" }
 wit-bindgen-teavm-java = { path = 'crates/teavm-java', version = '0.12.0' }
 wit-bindgen-go = { path = 'crates/go', version = '0.12.0' }
 wit-bindgen-markdown = { path = 'crates/markdown', version = '0.12.0' }
-wit-bindgen-rust-lib = { path = 'crates/rust-lib', version = '0.12.0' }
 wit-bindgen = { path = 'crates/guest-rust', version = '0.12.0', default-features = false }
 wit-bindgen-rust-macro-shared = { path = 'crates/rust-macro-shared', version = '0.3.0' }
