--- crates/rust-lib/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "wit-bindgen-rust-lib"
-authors = ["Alex Crichton <alex@alexcrichton.com>"]
-version = "0.12.0"
-edition.workspace = true
-repository = 'https://github.com/bytecodealliance/wit-bindgen'
-license = "Apache-2.0 WITH LLVM-exception"
-homepage = 'https://github.com/bytecodealliance/wit-bindgen'
-description = """
-Shared support for generating Rust code with wit-bindgen-style code generators.
-"""
-
-[lib]
-doctest = false
-test = false
-
-[dependencies]
-wit-bindgen-core = { workspace = true }
-heck = { workspace = true }
