+++ crates/rust-macro/Cargo.toml
@@ -21,6 +21,5 @@ syn = { version = "2.0", features = ["printing"] }
 quote = "1"
 wit-bindgen-core = { workspace = true }
 wit-bindgen-rust = { workspace = true }
-wit-bindgen-rust-lib = { workspace = true }
 wit-component = { workspace = true }
 anyhow = { workspace = true }
