+++ crates/rust/Cargo.toml
@@ -18,7 +18,6 @@ doctest = false
 [dependencies]
 anyhow = { workspace = true }
 wit-bindgen-core = { workspace = true }
-wit-bindgen-rust-lib = { workspace = true }
 wit-component = { workspace = true }
 wasm-metadata = { workspace = true }
 heck = { workspace = true }
